notebook/requirements.txt
weaviate-client>=3.26.7,<4.0.0
langchain==0.1.14
langchain-community==0.0.29
ollama==0.1.8
langchain-ollama==0.1.0
python-dotenv>=1.0.0  # если используете .env

